Create TERMS.md
I have created TERMS and Conditions for STEEMGIGS with the permission of Jalasem.
It is of utmost importance that the STEEMGIGS platform have TERMS and Conditions for it's users. 
This will help guide against infringement of other intellectual property and help guide users on the STEEMGIGS platform.

Please review and merge.

I would also love to be a renowned writer for STEEMGIGS by always being available to help cross check and ensure that write-ups for the STEEMGIGS platform are always up to date, if you don't mind.