Update from Task #153 - added new suggestions
While exploring results, user will be shown a link to create a custom request of micro-task if he fails to find his desired gig. When, no results gets found for a query, it gives you some bunch of options now instead of only "Explore Gigs". user will have a variety of options to start with when can't find his gig.